backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
ultralytics
requests
fpdf
reprtlab